+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changelog
 - Add @booking-move restapi
   [mamico]
 
+- Extend @@bookings search view parameters list.
+  [folix-01]
 - Added event handler on booking creation to send email to managers.
   [daniele]